pyproject.toml
[tool.poetry]
name = "learning_diary"
version = "0.1.0"
description = "Tool to keep track of studied stuff."
license = "MIT"
authors = [
    "1uy31"
]
readme = "README.md"  # Markdown files are supported

[tool.poetry.dependencies]
python = "3.10.7"  # Compatible python versions must be declared here
python-dotenv = "0.19.2"
Flask = "2.0.3"
Flask-Migrate = "3.1.0"
Flask-SQLAlchemy = "2.5.1"
SQLAlchemy = "^1.4.31"
psycopg2 = "2.9.3"
graphene = "2.1.9"
graphene-sqlalchemy = "^2.3.0"
Flask-GraphQL = "2.0.1"
pytz = "^2021.3"
types-toml = "^0.10.4"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
pylint = "^2.12.2"
mypy = "^0.931"
pytest = "^7.0.1"
factory-boy = "^3.2.1"
Faker = "^13.0.0"
coverage = "^6.3.1"
tox = "^3.24.5"
pre-commit = "2.17.0"

[tool.poetry.scripts]
